move Direction Button into route Component
